3rd increment
admin's 3rd increment

+++ admin/manage_hospital/application/delete_hospital.php
@@ -1,14 +1,29 @@
 <?php
 	require '../../session.php';
 
-	if (isset($_GET['delid'])) {
+	if (isset($_GET['delid']) && isset($_GET['name']) && isset($_GET['dis']) && isset($_GET['add']) ) {
         $hosid= $_GET['delid'];
-        $sql="DELETE FROM blood_bank_hospital WHERE HospitalID='$hosid'";
-        if ($result=mysqli_query($link, $sql)) {
+        $name= $_GET['name'];
+        $dis= $_GET['dis'];
+        $add= $_GET['add'];
+
+        $sql2="SELECT NIC FROM blood_bank_admin WHERE BloodBankID='$hosid'";
+        $sql3="SELECT DonateID FROM donate_hospital WHERE HospitalID='$hosid' UNION SELECT DonateID FROM donate_campaign WHERE HospitalID='$hosid'";
+        $result2=mysqli_query($link, $sql2);$result3=mysqli_query($link,$sql3);
+        if (mysqli_num_rows($result2)==0 && mysqli_num_rows($result3)==0) {
+            $sql="DELETE FROM blood_bank_hospital WHERE HospitalID='$hosid'";
+            if ($result=mysqli_query($link, $sql)) {
             header("Location: ../index?del=ok");
-        }else{
+            }else{
             echo "Can't Delete";
+            }
+        }elseif(mysqli_num_rows($result2)>0){
+            header("Location: ../delete_hospital?admin=pass&name=$name&dis=$dis&add=$add&id=$hosid");
+        }else{
+            header("Location: ../delete_hospital?packet=pass&name=$name&dis=$dis&add=$add&id=$hosid");
         }
+
+        
     }
 
     mysqli_close($link);
